requirements.txt
Django==4.0.6
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0
pytest==7.1.2
psycopg2==2.9.3
pytest-django==4.5.2
pytest-drf==1.1.3
flake8==4.0.1
